+++ src/localisation/chineseSimplified.ts
@@ -220,6 +220,8 @@ const CHINESE_SIMPLIFIED: Translations = {
   [Phrase.PlaybackSpeedTooltip]: '回放速度',
   [Phrase.ClipTooltip]: '剪辑',
   [Phrase.ClipUnavailableTooltip]: '只能剪辑本地保存的视频',
+  [Phrase.ClipSourceTooltip]: '转到源录制',
+  [Phrase.ClipSourceUnavailableTooltip]: '源录制不可用',
   [Phrase.ConfirmTooltip]: '确认',
   [Phrase.CancelTooltip]: '取消',
   [Phrase.TagButtonTooltip]: '添加标签',
@@ -560,6 +562,12 @@ const CHINESE_SIMPLIFIED: Translations = {
   [Phrase.CharacterFilterAddManual]: '...或手动添加角色筛选。',
   [Phrase.CharacterFilterNone]: '当前未启用角色上传筛选。添加角色以启用该筛选。',
   [Phrase.CharacterFilterActive]: '仅会上传使用下方所列角色进行游戏时创建的录制。',
+  [Phrase.Arrows]: '箭头',
+  [Phrase.Navigate]: '导航',
+  [Phrase.ClippedAtLabel]: '截断于',
+  [Phrase.SelectRange]: '选择范围',
+  [Phrase.SelectMultiple]: '选择多个',
+  [Phrase.SelectAll]: '全选',
 };
 
 export default CHINESE_SIMPLIFIED;