+++ src/main/Manager.ts
@@ -630,7 +630,14 @@ export default class Manager {
 
         const sourceMetadata = rendererVideoToMetadata({ ...video });
         const now = new Date();
-        const clipMetadata = buildClipMetadata(sourceMetadata, duration, now);
+
+        const clipMetadata = buildClipMetadata(
+          sourceMetadata,
+          duration,
+          now,
+          offset,
+          video.videoName,
+        );
 
         const clipQueueItem: VideoQueueItem = {
           name: video.videoName,