+++ src/parsing/LogHandler.ts
@@ -54,24 +54,38 @@ export default abstract class LogHandler {
   private static stateChangeCallback: () => void;
 
   /**
+   * The timer is static and thus shared across all LogHandlers. We don't want
+   * any weird behaviour when quickly switching between games. See Issue 855.
+   */
+  private static logDataTimeout: NodeJS.Timeout | null = null;
+
+  /**
+   * The lower this timer the better, but log flush behaviour varies between
+   * games and we can't make it too low or we risk ending activities early. Let
+   * the specific log handler instance decide what the timeout should be.
+   */
+  private logDataTimeoutMs: number;
+
+  /**
    * Enforces ordered processing of log lines. Some log line processing
    * is asynchronous so we need to ensure later lines don't get processed
    * before earlier ones.
    */
   protected logProcessQueue = new AsyncQueue(Number.MAX_SAFE_INTEGER);
 
-  constructor(logPath: string, dataTimeout: number) {
-    this.combatLogWatcher = new CombatLogWatcher(logPath, dataTimeout);
+  constructor(logPath: string, dataTimeoutMins: number) {
+    this.logDataTimeoutMs = dataTimeoutMins * 60 * 1000;
+    this.combatLogWatcher = new CombatLogWatcher(logPath);
     this.combatLogWatcher.watch();
     const lpq = this.logProcessQueue;
-
-    this.combatLogWatcher.on('timeout', (ms) => {
-      lpq.add(async () => this.dataTimeout(ms));
-    });
 
     // For ease of testing force stop.
     this.combatLogWatcher.on('WARCRAFT_RECORDER_FORCE_STOP', () => {
       lpq.add(async () => LogHandler.forceEndActivity());
+    });
+
+    this.combatLogWatcher.on('WARCRAFT_RECORDER_LOG_ACTIVITY', () => {
+      lpq.add(async () => this.resetTimeout());
     });
   }
 
@@ -218,9 +232,7 @@ export default abstract class LogHandler {
       return;
     }
 
-    console.info(
-      `[LogHandler] Start recording a video for category: ${category}`,
-    );
+    console.info(`[LogHandler] Start an activity for category: ${category}`);
 
     // Offset is the number of seconds to cut back into the buffer. That way
     // the buffer length is irrelevant. It is physically impossible to have
@@ -360,14 +372,28 @@ export default abstract class LogHandler {
 
   protected async dataTimeout(ms: number) {
     console.info(
-      `[LogHandler] Haven't received data for combatlog in ${
+      `[LogHandler] Haven't received data from any combat logs in ${
         ms / 1000
       } seconds.`,
     );
 
-    if (LogHandler.activity) {
-      await LogHandler.forceEndActivity(-ms / 1000);
+    if (!LogHandler.activity) {
+      console.info('[LogHandler] No activity, no action');
+      return;
     }
+
+    if (LogHandler.overrunning) {
+      console.info('[LogHandler] Activity in overrun, no action');
+      return;
+    }
+
+    if (this.isManual()) {
+      console.info('[LogHandler] Manual recording, no action');
+      return;
+    }
+
+    console.info('[LogHandler] Force ending activity due to data timeout');
+    await LogHandler.forceEndActivity(-ms / 1000);
   }
 
   public static async forceEndActivity(timedelta = 0) {
@@ -537,5 +563,21 @@ export default abstract class LogHandler {
 
     console.warn('[LogHandler] Unable to start manual recording');
     if (sounds) playSoundAlert(SoundAlerts.MANUAL_RECORDING_ERROR);
+  }
+
+  /**
+   * Reset the log data timeout. This should be called whenever we receive
+   * data from the combat log.
+   */
+  protected resetTimeout() {
+    if (LogHandler.logDataTimeout) {
+      clearTimeout(LogHandler.logDataTimeout);
+    }
+
+    LogHandler.logDataTimeout = setTimeout(() => {
+      this.logProcessQueue.add(async () =>
+        this.dataTimeout(this.logDataTimeoutMs),
+      );
+    }, this.logDataTimeoutMs);
   }
 }