+++ src/renderer/rendererutils.ts
@@ -37,6 +37,7 @@ import {
   AudioSource,
   AppState,
   AudioSourceType,
+  RendererClip,
 } from 'main/types';
 import { ambiguate } from 'parsing/logutils';
 import { VideoCategory } from 'types/VideoCategory';
@@ -322,22 +323,24 @@ const getEncounterNameById = (encounterId: number): string => {
 };
 
 /**
- * Get the dungeon name if possible, else an empty string.
+ * Get the dungeon name if possible, else "Unknown Map".
  */
-const getDungeonName = (video: RendererVideo) => {
+const getDungeonName = (video: RendererVideo): string => {
   const { mapID } = video;
 
   if (mapID === undefined) {
-    return '';
+    return 'Unknown Map';
   }
 
   if (video.flavour === Flavour.Retail) {
-    return dungeonsByMapId[mapID];
+    return dungeonsByMapId[mapID] ?? 'Unknown Map';
   }
 
   if (video.flavour === Flavour.Classic) {
-    return mopChallengeModes[mapID];
+    return mopChallengeModes[mapID] ?? 'Unknown Map';
   }
+
+  return 'Unknown Map';
 };
 
 const isMythicPlusUtil = (video: RendererVideo) => {
@@ -1184,6 +1187,27 @@ const formatRealmNameForDisplay = (realm: string) => {
   return realm.replace(/([A-Z])|(?<!\d)(\d)/g, ' $1$2').trim();
 };
 
+const findClipParent = (
+  clip: RendererClip,
+  videos: RendererVideo[],
+): RendererVideo | undefined => {
+  if (clip.player?._GUID === 'WCR MultiPov GUID') {
+    // Multiview videos are clips and we might be able to handle them but I
+    // can't be bothered with the headache for now.
+    return undefined;
+  }
+
+  const parentVideoName = clip.parentVideoName
+    ? clip.parentVideoName
+    : clip.videoName.slice(0, -33); // " - Clipped at ..." is 33 characters.
+
+  const parent = videos.find(
+    (candidate) => candidate.videoName === parentVideoName,
+  );
+
+  return parent;
+};
+
 export {
   getFormattedDuration,
   getVideoResult,
@@ -1251,4 +1275,5 @@ export {
   getFriendlyCodecName,
   isHevcEncoder,
   formatRealmNameForDisplay,
+  findClipParent,
 };